More modelling and pairplot
Changed focus of analysis to look at values and not units adj.
Added in MAE as an analysis of the metrics
Added in ARIMA Model. Not great compared to KNN. by far the best model and to include all the variables by sacrificing our R2 to maintain the integrity of the data.